Revert to Stable
The network swap is unstable right now, so its being developed in a
separate branch. It does not interfere with ingame development though so
don't let that hold you back :D Since the changes are quite large it
would break source changes so current development of this state source
is not recommended.